requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
openai
passlib[bcrypt]
python-jose
python-multipart
httpx
pytest
pycocotools
ultralytics>=8.3.0
opencv-python-headless
git+https://github.com/facebookresearch/sam2.git
torch==2.6.0+cu124
torchvision==0.21.0+cu124
prometheus_client
starlette-exporter
python-json-logger
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-psycopg2
sentry-sdk[fastapi]  
sentry-sdk[sqlalchemy] 
sentry-sdk[logging]
sentry-cli 
pillow
groundingdino-py
addict
supervision
timm
yapf
--extra-index-url https://download.pytorch.org/whl/cu124
